union hierarchical and flat structure in PropertiesConfigProcessor.
adding new config attribute 'hierarchical'

change docs&test
